Use go-flags env var parsing; display HTTP rcode
Removed the explicit parsing of OS environment variables for Access Key
and Secret Key and instead rely on the ability of the go-flags library
to do it using the env parameter.

Also updated the response output so that HTTP response codes are printed
if they are 300 or greater. This eliminates completely empty output for
responses such as 401 Unauthorized.